Added a simple data versioning strategy.
You can now specify a `version` field when creating a Hoard.  This version is simply an opaque string.  Youcan also specify a `versionChangeHandler` which gets called when the passed-in version does not match a saved version from a previous instance of that namespace.

You can implement a version change handling strategy of your choice in the `versionChangeHandler`, e.g. blowing away all records or migrating them to the new data schema.  Returning `true` from `versionChangeHandler` tells Nibelung that you've handled the change and causes it to stamp the namespace with the new version.

Supplied a default change handler that does nothing, for backwards compatibility.